pyproject.toml
[tool.poetry]
name = "f1-backend"
version = "0.1.0"
description = ""
authors = ["JoshMusira <joshuamusira01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["standard"], version = "^0.115.12"}
uvicorn = {extras = ["standard"], version = "^0.34.0"}
python-dotenv = "^1.1.0"
langchain-community = "^0.3.20"
langchain = "^0.3.21"
langchain-openai = "^0.3.11"
uuid = "^1.30"
langdetect = "^1.0.9"
pinecone = {extras = ["grpc"], version = "^6.0.2"}
langchain-google-genai = "^2.1.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
